Спринт 15
Clean Architecture
Попытка 3.1 Всё работает.
 Добавлен enum-класс состояний mediaPlayer.
 Убран мусор.
