+++ app/src/main/java/com/practicum/playlistmaker/TrackItemAdapter.kt
@@ -8,7 +8,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.google.gson.Gson
-import com.practicum.playlistmaker.player.presentation.PlayerActivity
+import com.practicum.playlistmaker.playerActivity.presentation.PlayerActivity
 
 var isClickAllowed = true
