--- app/src/main/java/com/practicum/playlistmaker/player/domain/interfaces/TrackInteractor.kt
@@ -1,12 +1,0 @@
-package com.practicum.playlistmaker.player.domain.interfaces
-
-import com.practicum.playlistmaker.player.domain.entities.Track
-
-interface TrackInteractor {
-    fun getTrack(): Track
-
-//    fun preparePlayer()
-    fun startPlayer(): Int
-    fun pausePlayer(): Int
-    fun playbackControl(): Int
-}
