--- app/src/main/java/com/practicum/playlistmaker/player/domain/interfaces/TrackRepository.kt
@@ -1,16 +1,0 @@
-package com.practicum.playlistmaker.player.domain.interfaces
-
-import com.practicum.playlistmaker.player.domain.entities.Track
-
-interface TrackRepository {
-    fun getTrack(): Track
-
-    fun preparePlayer(mediaPlayerPreparator:MediaPlayerPrepare)
-
-    fun startPlayer():Int
-
-    fun pausePlayer():Int
-
-//    fun setOnCompletionListener(onCompletion: () -> Unit)
-
-}
