+++ app/src/main/java/com/practicum/playlistmaker/playerActivity/data/intent/TrackIntentDAOImpl.kt
@@ -1,9 +1,9 @@
-package com.practicum.playlistmaker.player.data.intent
+package com.practicum.playlistmaker.playerActivity.data.intent
 
 import android.content.Intent
 import com.google.gson.Gson
-import com.practicum.playlistmaker.player.domain.dao.TrackIntentDAO
-import com.practicum.playlistmaker.player.domain.entities.Track
+import com.practicum.playlistmaker.playerActivity.domain.dao.TrackIntentDAO
+import com.practicum.playlistmaker.playerActivity.domain.entities.Track
 
 
 class TrackIntentDAOImpl(val intent: Intent): TrackIntentDAO {
