+++ app/src/main/java/com/practicum/playlistmaker/playerActivity/domain/dao/TrackIntentDAO.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.playerActivity.domain.dao
+
+import com.practicum.playlistmaker.playerActivity.domain.entities.Track
+
+interface TrackIntentDAO {
+    fun getTrack():Track
+}
