+++ app/src/main/java/com/practicum/playlistmaker/playerActivity/domain/entities/MediaPlayerState.kt
@@ -1,0 +1,8 @@
+package com.practicum.playlistmaker.playerActivity.domain.entities
+
+enum class MediaPlayerState {
+    STATE_DEFAULT,
+    STATE_PREPARED,
+    STATE_PLAYING,
+    STATE_PAUSED,
+}
