+++ app/src/main/java/com/practicum/playlistmaker/playerActivity/domain/usecases/TrackInteractorImlp.kt
@@ -1,0 +1,35 @@
+package com.practicum.playlistmaker.playerActivity.domain.usecases
+
+import com.practicum.playlistmaker.playerActivity.domain.entities.MediaPlayerState
+import com.practicum.playlistmaker.playerActivity.domain.entities.Track
+import com.practicum.playlistmaker.playerActivity.domain.interfaces.TrackInteractor
+import com.practicum.playlistmaker.playerActivity.domain.interfaces.TrackRepository
+
+class TrackInteractorImlp(
+    private val trackRepository: TrackRepository,
+    var playerState: MediaPlayerState
+) : TrackInteractor {
+    override fun playbackControl(): MediaPlayerState {
+        when (playerState) {
+            MediaPlayerState.STATE_PLAYING -> playerState = pausePlayer()
+            MediaPlayerState.STATE_PAUSED, MediaPlayerState.STATE_PREPARED -> playerState =
+                startPlayer()
+            else -> {}
+        }
+        return playerState
+    }
+
+    override fun getTrack(): Track {
+        return trackRepository.getTrack()
+    }
+
+    override fun startPlayer(): MediaPlayerState {
+        playerState = trackRepository.startPlayer()
+        return playerState
+    }
+
+    override fun pausePlayer(): MediaPlayerState {
+        playerState = trackRepository.pausePlayer()
+        return playerState
+    }
+}
